requirements.txt
pandas
gspread
oauth2client
google-generativeai
requests
pytz